moved the check
