+++ app/notifications/process_notifications.py
@@ -8,6 +8,7 @@ from notifications_utils.recipients import (
     get_international_phone_info,
     validate_and_format_phone_number,
 )
+from notifications_utils.template import SMSMessageTemplate
 from notifications_utils.timezones import convert_local_timezone_to_utc
 
 from app import redis_store
@@ -28,12 +29,25 @@ from app.models import (
     LETTER_TYPE,
     NOTIFICATION_CREATED,
     SMS_TYPE,
+    ApiKeyType,
     Notification,
     ScheduledNotification,
+    Service,
 )
-from app.sms_fragment_utils import increment_daily_sms_fragment_count
+from app.sms_fragment_utils import (
+    fetch_daily_sms_fragment_count,
+    increment_daily_sms_fragment_count,
+)
 from app.utils import get_template_instance
-from app.v2.errors import BadRequestError
+from app.v2.errors import BadRequestError, LiveServiceTooManySMSRequestsError
+
+
+def check_if_request_would_put_service_over_daily_sms_limit(key_type: ApiKeyType, service: Service, requested_sms: int):
+    if current_app.config["FF_SPIKE_SMS_DAILY_LIMIT"] and key_type != KEY_TYPE_TEST and current_app.config["REDIS_ENABLED"]:
+        fragments_sent = fetch_daily_sms_fragment_count(service.id)
+        total_requested = fragments_sent + requested_sms
+        if total_requested > service.sms_daily_limit:
+            raise LiveServiceTooManySMSRequestsError(service.sms_daily_limit)
 
 
 def create_content_for_notification(template, personalisation):
@@ -63,11 +77,11 @@ def persist_notification(
     template_id,
     template_version,
     recipient,
-    service,
+    service: Service,
     personalisation,
     notification_type,
     api_key_id,
-    key_type,
+    key_type: ApiKeyType,
     created_at=None,
     job_id=None,
     job_row_number=None,
@@ -129,6 +143,15 @@ def persist_notification(
 
     # if simulated create a Notification model to return but do not persist the Notification to the dB
     if not simulated:
+        if notification_type == SMS_TYPE:
+            _template = SMSMessageTemplate(
+                template.__dict__,
+                values=notification.personalisation,
+                prefix=service.name,
+                show_prefix=service.prefix_sms,
+            )
+            check_if_request_would_put_service_over_daily_sms_limit(key_type, service, _template.fragment_count)
+
         dao_create_notification(notification)
         if key_type != KEY_TYPE_TEST:
             if redis_store.get(redis.daily_limit_cache_key(service.id)):
