+++ app/notifications/rest.py
@@ -20,7 +20,6 @@ from app.notifications.process_notifications import (
     simulated_recipient,
 )
 from app.notifications.validators import (
-    check_if_request_would_put_service_over_daily_sms_limit,
     check_rate_limiting,
     check_template_is_active,
     check_template_is_for_notification_type,
@@ -137,11 +136,6 @@ def send_notification(notification_type: NotificationType):
         reply_to_text=template.get_reply_to_text(),
     )
     if not simulated:
-        if notification_type == SMS_TYPE:
-            check_if_request_would_put_service_over_daily_sms_limit(
-                api_user.key_type, authenticated_service, notification_model.billable_units
-            )
-
         send_notification_to_queue(
             notification=notification_model,
             research_mode=authenticated_service.research_mode,
